refactor: 뮤테이션 및 쿼리 분리 및 리뷰사항 반영
